requirements.txt
openai
numpy
pillow
nuscenes-devkit
scikit-learn
python-dotenv
tqdm